% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_cdr3_dist.R
\name{aa_cdr3_dist}
\alias{aa_cdr3_dist}
\title{Average frequency of nucleotide mutations resulting in amino acid substitution in CDR3}
\usage{
aa_cdr3_dist(data)
}
\arguments{
\item{data}{A column CDR3_IMGT from IMGT/HighV-Quest table 7 called "7_V-REGION-mutation-and-AA-change-table.txt"}
}
\value{
Output is the data matrix of average frequency of nucleotide mutations resulting in amino acid substitution.
}
\description{
Calculate average frequency of nucleotide mutations resulting in amino acid substitution in CDR3.
}
\examples{
data(IMGTtab7)
cdr3_matrix<-aa_cdr3_dist(data=IMGTtab7$CDR3_IMGT)

}
\references{
Zuckerman NS., Hazanov H., Barak M., Edelman H., Hess S., Shcolnik H., Dunn-Walters D.,and Mehr R. Somatic hypermutation and antigen-driven selection of B cells are altered inautoimmune diseases.J Autoimmun, 35(4):325 - 335, 2010. doi: 10.1016/j.jaut.2010.07.004.
}
