% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_comb_funk.R
\name{gene_comb_funk}
\alias{gene_comb_funk}
\title{gene/gene ratios of two gene families}
\usage{
gene_comb_funk(family1, family2)
}
\arguments{
\item{family1}{"V_GENE_and_allele" colums from IMGT/HighV-Quest table 1 coled "1_Summary.txt"}

\item{family2}{"J_GENE_and_allele" or "D_GENE_and_allele" colums from IMGT/HighV-Quest table 1 coled "1_Summary.txt"}
}
\value{
Output is the data matrix of gene/gene ratios of two gene families.
}
\description{
The calculation was obtained according bcRep package in R language. Every combination of two families was summarized and divided by the sum of all combinations across each group.
}
\examples{
data(IMGTtab1)
gane_comb<-gene_comb_funk(family1 = IMGTtab1$V_GENE_and_allele,
                          family2 = IMGTtab1$J_GENE_and_allele)
gane_comb

}
\references{
Bischof J. and Ibrahim SM. bcRep: R Package for Comprehensive Analysis of B Cell Receptor Repertoire Data. PLoS One. 11(8):e0161569, 2016. doi: 10.1371/journal.pone.0161569.
}
