% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/germlineReconstr.R
\name{germlineReconstr}
\alias{germlineReconstr}
\title{Function to reconstruction of germline sequencies based on IMGT outputs.}
\usage{
germlineReconstr(data_tab2, data_tab7)
}
\arguments{
\item{data_tab2}{Column V_Region in the IMGT table 2}

\item{data_tab7}{Column V_Region in the IMGT table 7}
}
\value{
output is a vector of germline sequencies.
}
\description{
Function to reconstruction of germline sequencies based on IMGT outputs.
}
\examples{
data(IMGTtab2)
data(IMGTtab7)
germline<-germlineReconstr(IMGTtab2$V_REGION, IMGTtab7$V_REGION)
germline
}
\references{
Brochet X., Lefranc MP., and Giudicelli V. IMGT/V-QUEST: the highly customized andintegrated system for IG and TR standardized V-Jand V-D-J sequence analysis.NucleicAcids Res., 36(Web Server issue):W503 - W508, 2008. doi: 10.1093/nar/gkn316
}
