% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTtab1}
\alias{IMGTtab1}
\title{IMGT Table 1}
\format{A data frame with 9 rows and 29 variables}
\usage{
data(IMGTtab1)
}
\description{
Data from the Excel file 1 are named Summary and are included in IMGTtab1
}
\details{
\itemize{
  \item Sequence_number. number of Sequence
  \item Sequence_ID. ID frome gene bank
  \item Functionality. productive/unproductive
  \item V_GENE_and_allele. germline genes of V Region
  \item J_GENE_and_allele. germline genes of J Region
  \item D_GENE_and_allele. germline genes of D Region
  }
}
\keyword{datasets}
