% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTtab2}
\alias{IMGTtab2}
\title{IMGT Table 2}
\format{A data frame with 9 rows and 18 variables}
\usage{
data(IMGTtab2)
}
\description{
Data from the Excel file 2 are named IMGT-gapped-nt-sequences and are included in IMGTtab2
}
\details{
\itemize{
  \item Sequence_number. number of Sequence
  \item Sequence_ID. ID frome gene bank
  \item Functionality. productive/unproductive
  \item V_GENE_and_allele. germline genes of V Region
  \item J_GENE_and_allele. germline genes of J Region
  \item D_GENE_and_allele. germline genes of D Region
  \item V_D_J_REGION. nucleotides in VDJ region in mutated sequence
  \item V_J_REGION. NA
  \item V_REGION. nucleotides in V region in mutated sequence
  \item FR1_IMGT. nucleotides in FR1 in mutated sequence
  \item CDR1_IMGT. nucleotides in CDR1 in mutated sequence
  \item FR2_IMGT. nucleotides in FR2 in mutated sequence
  \item CDR2_IMGT. nucleotides in CDR2 in mutated sequence
  \item FR3_IMGT. nucleotides in FR3 in mutated sequence
  \item CDR3_IMGT. nucleotides in CDR3 in mutated sequence
  \item JUNCTION. nucleotides in JUNCTION region in mutated sequence
  \item J_REGION. nucleotides in J region in mutated sequence
  \item FR4_IMGT. nucleotides in FR4 in mutated sequence
  }
}
\keyword{datasets}
