% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTtab7}
\alias{IMGTtab7}
\title{IMGT Table 7}
\format{A data frame with 9 rows and 11 variables}
\usage{
data(IMGTtab7)
}
\description{
A dataset containing the output from IMGT/HighV-QUEST
 IMGT Table 7. The variables are as follows:
}
\details{
\itemize{
  \item Sequence_number. number of Sequence
  \item Sequence_ID. ID frome gene bank
  \item Functionality. productive/unproductive
  \item V_GENE_and_allele. germline genes of V Region
  \item V_REGION. nucleotide changes and aminoacid changes in V region
  \item FR1_IMGT. nucleotide changes and aminoacid changes in FR1
  \item CDR1_IMGT. nucleotide changes and aminoacid changes in CDR1
  \item FR2_IMGT. nucleotide changes and aminoacid changes in FR2
  \item  CDR2_IMGT. nucleotide changes and aminoacid changes in CDR2
  \item  FR3_IMGT. nucleotide changes and aminoacid changes in FR3
  \item CDR3_IMGT. nucleotide changes and aminoacid changes in CDR3
  }
}
\keyword{datasets}

