% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Klassen}
\alias{Klassen}
\title{Klassen}
\format{A data frame with 400 rows and 2 variables}
\usage{
data(Klassen)
}
\description{
A dataset for usage of Funktion aa_plotgg_skal()
 Klassen. The variables are as follows:
}
\details{
\itemize{
  \item Egenschaften. Physio-chemical properties of aminoacids
  \item Klassen. Classification in classes according to IMGT/HighV-QUEST classes
  }
}
\keyword{datasets}

