% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_funktion.R
\name{na_funktion}
\alias{na_funktion}
\title{The function to check for reconsrtruction errors}
\usage{
na_funktion(data)
}
\arguments{
\item{data}{from function germlineReconstr().}
}
\value{
total number of NAs. If the answer is null then there are no errors in the germline reconstruction.
}
\description{
The function to check for reconsrtruction errors
}
\examples{
germline<-germlineReconstr(IMGTtab2$V_REGION, IMGTtab7$V_REGION)
na_funktion(germline)

}
