% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IMGTtab8}
\alias{IMGTtab8}
\title{IMGT Table 8}
\format{A data frame with 9 rows and 130 variables}
\usage{
data(IMGTtab8)
}
\description{
A dataset containing the output from IMGT/HighV-QUEST
 IMGT Table 8. The variables are as follows:
}
\details{
\itemize{
  \item Sequence_number. number of Sequence
  \item Sequence_ID. ID frome gene bank
  \item Functionality. productive/unproductive
  \item V_GENE_and_allele. germline genes of V Region
  }
}
\keyword{datasets}

